+++ examples/args.c
@@ -8,24 +8,33 @@ int main(int argc, char **argv) {
     // All options take: args*, long name, description, and zero or more named arguments.
     // Named arguments are set through designated initializer: `.name1 = value1, .name2 = value2`,
     // with default value 0 / '\0' / false / NULL depending on the type.
-    // Named arguments: short_name, required, default_value.
+    // Named arguments: short_name, required, hidden, default_value.
     const bool *help = option_flag(&a, "help", "Show help", .short_name = 'h', .ignore_required = true);
     const bool *version = option_flag(&a, "version", "Print version", .short_name = 'v', .ignore_required = true);
-    const long *l = option_long(&a, "long", "A long option", .short_name = 'l', .required = false);
-    const float *f = option_float(&a, "float", "A float option", .short_name = 'f', .required = true);
-    const char **s = option_string(&a, "str", "A string option", .short_name = 's', .default_value = "string default");
-    const char **p = option_path(&a, "path", "A path option");
+    const bool *hidden = option_flag(&a, "hidden", NULL, .hidden = true);
+    const long *long_ = option_long(&a, "long", "A long option", .short_name = 'l');
+    const float *float_ = option_float(&a, "float", "A float option", .short_name = 'f', .required = true);
+    const char **string = option_string(
+        &a,
+        "str",
+        "A string option",
+        .short_name = 's',
+        .default_value = "string default"
+    );
+    const char **path = option_path(&a, "path", "A path option");
 
     // If enum is continuous and array matches it, result of `option_enum` can be converted directly.
     typedef enum { FIRST, SECOND, THIRD } Enum;
     const char *enum_values[] = {"first", "second", "third", NULL};
-    const Enum *e = (const Enum *) option_enum(
-        &a, "enum", "An enum option", enum_values, .short_name = 'e', .default_value = FIRST
-    );
+    const Enum *enum_ = (const Enum *)
+        option_enum(&a, "enum", "An enum option", enum_values, .short_name = 'e', .default_value = FIRST);
     // If values don't match, or the enum isn't continuous, it may be desirable to get a string instead.
     // Also, if values array are defined in the arguments, it must be wrapped in parenthesis.
-    const char **es = option_enum_string(
-        &a, "enum-str", "A string enum option", ((const char *[]) {"other", "enum", "values", NULL}),
+    const char **enum_string = option_enum_string(
+        &a,
+        "enum-str",
+        "A string enum option",
+        ((const char *[]) {"other", "enum", "values", NULL}),
         .default_value = "default"
     );
 
@@ -60,7 +69,16 @@ int main(int argc, char **argv) {
     }
 
     // Use option values.
-    printf("options: l=%ld f=%f s=\"%s\" p=\"%s\" e=%d E=\"%s\"\n", *l, *f, *s, *p, *e, *es);
+    printf(
+        "options: h=%s l=%ld f=%f s=\"%s\" p=\"%s\" e=%d E=\"%s\"\n",
+        *hidden ? "true" : "false",
+        *long_,
+        *float_,
+        *string,
+        *path,
+        *enum_,
+        *enum_string
+    );
 
     // Handle positional arguments.
     printf("positional arguments:");